[refactor]: перенос лэйаутов по FSD, рефакторинг футера, изоляция роутов заглушек и исправление вложенности логотипа #792
