+++ frontend/src/app/router.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router'
-import { PublicLayout, AppLayout, AdminLayout } from '@widgets/layout'
+import { PublicLayout, AppLayout, AdminLayout } from './layouts'
 import { Landing } from '@pages/public/Landing'
 import { Login } from '@pages/public/Login'
 import { Register } from '@pages/public/Register'
@@ -156,6 +156,10 @@ export const router = createBrowserRouter([
           '1202',
         ),
       },
+      {
+        path: '*',
+        element: <NotFound />,
+      },
     ],
   },
   {
@@ -201,6 +205,10 @@ export const router = createBrowserRouter([
           'Управление источниками вакансий.',
           '1197',
         ),
+      },
+      {
+        path: '*',
+        element: <NotFound />,
       },
     ],
   },
