--- frontend/src/widgets/layout/index.ts
@@ -1,3 +1,0 @@
-export { AppLayout } from './ui/AppLayout'
-export { PublicLayout } from './ui/PublicLayout'
-export { AdminLayout } from './ui/AdminLayout'
