+++ frontend/src/widgets/sidebar/ui/Sidebar.tsx
